v1.0.2 changes 1. Updated README.md 2. Default backup interval is 600 seconds 3. Refactored ThreadRecurringTimer to RecurringTimer 4. Unittest case; Backup should only be tested for python version >= 3.7

+++ sqlite_rx/tests/backup/conftest.py
@@ -1,4 +1,5 @@
 import os
+import sys
 import platform
 import signal
 import tempfile
@@ -21,32 +22,34 @@ backup_event = namedtuple('backup_event', ('client', 'backup_database'))
 
 @pytest.fixture(scope="module")
 def plain_client():
-    auth_config = {
-        sqlite3.SQLITE_OK: {
-            sqlite3.SQLITE_DROP_TABLE
+    
+    if (sys.version_info.major == 3 and sys.version_info.minor >= 7):
+        auth_config = {
+            sqlite3.SQLITE_OK: {
+                sqlite3.SQLITE_DROP_TABLE
+            }
         }
-    }
-    with tempfile.NamedTemporaryFile() as back_db_file, tempfile.NamedTemporaryFile() as main_db_file:
-        server = SQLiteServer(bind_address="tcp://127.0.0.1:5003",
-                              database=main_db_file.name,
-                              auth_config=auth_config,
-                              backup_database=back_db_file.name,
-                              backup_interval=1)
-        
-        client = SQLiteClient(connect_address="tcp://127.0.0.1:5003")
+        with tempfile.NamedTemporaryFile() as back_db_file, tempfile.NamedTemporaryFile() as main_db_file:
+            server = SQLiteServer(bind_address="tcp://127.0.0.1:5003",
+                                database=main_db_file.name,
+                                auth_config=auth_config,
+                                backup_database=back_db_file.name,
+                                backup_interval=1)
+            
+            client = SQLiteClient(connect_address="tcp://127.0.0.1:5003")
 
-        event = backup_event(client=client, backup_database=back_db_file.name)
+            event = backup_event(client=client, backup_database=back_db_file.name)
 
-        server.start()
+            server.start()
 
-        LOG.info("Started Test SQLiteServer")
+            LOG.info("Started Test SQLiteServer")
 
-        yield event
+            yield event
 
-        if platform.system().lower() == 'windows':
-            os.system("taskkill  /F /pid "+str(server.pid))
-        else:
-            os.kill(server.pid, signal.SIGINT)
+            if platform.system().lower() == 'windows':
+                os.system("taskkill  /F /pid "+str(server.pid))
+            else:
+                os.kill(server.pid, signal.SIGINT)
 
-        server.join()
-        client.shutdown()
+            server.join()
+            client.shutdown()
